Working on a1.3
Function works (drawn by the 4 given points/draggers), but isn't rendered correctly. That's so because the calculatet points are not in the correct order in the points array.
The other problem is, if I drag one dragger, the curve is drawn new, but the old ones still exists. So in just one second there are hundreds of curves. This must be fixed, too.